Refactor Fireblocks CW registry and mapper structure
