+++ src/providers/fireblocks/cw/core/fireblocks-core.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { FireblocksClientProvider } from './providers/fireblocks-client.provider';
-import { FireblocksErrorMapper } from './providers/fireblocks-error-mapper';
+import { FireblocksErrorMapper } from '../infrastructure/persistence/relational/mappers/fireblocks-error.mapper';
 import { FireblocksResilience } from './providers/fireblocks-resilience';
 
 @Module({
